# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Errotel read API",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/capabilities": {
      "get": {
        "operationId": "getCapabilities",
        "responses": {
          "200": {
            "description": "Capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            },
            "headers": {
              "X-Errotel-Cache": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Cache-Age-Ms": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/occurrences/search": {
      "post": {
        "operationId": "searchOccurrences",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OccurrenceSearch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SearchResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "headers": {
              "X-Errotel-Cache": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Cache-Age-Ms": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/occurrences/histogram": {
      "post": {
        "operationId": "getHistogram",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OccurrenceSearch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HistogramResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistogramResponse"
                }
              }
            },
            "headers": {
              "X-Errotel-Cache": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Cache-Age-Ms": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/occurrences/{ref}": {
      "get": {
        "operationId": "getOccurrence",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OccurrenceDetail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrenceDetail"
                }
              }
            },
            "headers": {
              "X-Errotel-Cache": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Cache-Age-Ms": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/occurrences/{ref}/related": {
      "post": {
        "operationId": "getRelatedLogs",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelatedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "RelatedResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelatedResponse"
                }
              }
            },
            "headers": {
              "X-Errotel-Cache": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Cache-Age-Ms": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/traces/{traceId}": {
      "get": {
        "operationId": "getTrace",
        "parameters": [
          {
            "name": "traceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TraceResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceResponse"
                }
              }
            },
            "headers": {
              "X-Errotel-Cache": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Fetched-At": {
                "schema": {
                  "type": "string"
                }
              },
              "X-Errotel-Cache-Age-Ms": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "ApiError",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Wire_unixNano": {
        "type": "string",
        "pattern": "^(0|[1-9][0-9]*)$"
      },
...
      "TimeRange": {
        "type": "object",
        "properties": {
          "startUnixNano": {
            "$ref": "#/components/schemas/Wire_unixNano"
          },
          "endUnixNano": {
            "$ref": "#/components/schemas/Wire_unixNano"
          }
        },
        "required": [
          "startUnixNano",
          "endUnixNano"
        ],
        "additionalProperties": false,
        "description": "Absolute UTC Unix nanoseconds encoded as decimal strings. Start inclusive, end exclusive. Use capabilities.serverTimeUnixNano as a clock reference; respect maxRangeMs."
      },
      "SearchFilter": {
        "type": "object",
        "properties": {
          "op": {
            "type": "string",
            "enum": [
              "and",
              "or",
              "not",
              "eq",
              "contains",
              "icontains",
              "prefix",
              "exists"
            ]
          },
          "field": {
            "type": "string",
            "enum": [
              "service",
              "environment",
              "exceptionType",
              "message",
              "release",
              "traceId",
              "runtimeId",
              "groupKey",
              "origin"
            ]
          },
          "value": {
            "type": "string",
            "maxLength": 2048
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchFilter"
            },
            "maxItems": 64
          }
        },
        "required": [
          "op"
        ],
        "additionalProperties": false,
        "description": "Allowlisted AST, never raw LogsQL. and/or require nonempty children; not requires one child. Leaf operators require field and value (except exists). Use explicit grouping."
      },
      "OccurrenceSearch": {
        "type": "object",
        "properties": {
          "range": {
            "$ref": "#/components/schemas/TimeRange"
          },
          "service": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "exceptionType": {
            "type": "string"
          },
          "messageContains": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "runtimeId": {
            "type": "string"
          },
          "groupKey": {
            "type": "string"
          },
          "origin": {
            "type": "string",
            "enum": [
              "sdk",
              "otel-log",
              "both"
            ]
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "cursor": {
            "type": "string"
          },
          "filter": {
            "$ref": "#/components/schemas/SearchFilter"
          }
        },
        "required": [
          "range"
        ],
        "additionalProperties": false
      },
      "OccurrenceSummary": {
        "type": "object",
        "properties": {
          "ref": {
            "type": "string"
          },
          "origin": {
            "type": "string",
            "enum": [
              "sdk",
              "otel-log"
            ]
          },
          "eventId": {
            "type": "string"
          },
          "timestampUnixNano": {
            "$ref": "#/components/schemas/Wire_unixNano"
          },
          "service": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "release": {
            "type": "string"
          },
          "exceptionType": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "severityNumber": {
            "type": "integer"
          },
          "traceId": {
            "type": "string"
          },
          "spanId": {
            "type": "string"
          },
          "runtimeId": {
            "type": "string"
          },
          "contextStatus": {
            "type": "string",
            "enum": [
              "not_loaded",
              "absent",
              "invalid",
              "unsupported_version"
            ]
          }
        },
        "required": [
          "ref",
          "origin",
          "timestampUnixNano",
          "contextStatus"
        ],
        "additionalProperties": false
      },
      "Meta": {
        "type": "object",
        "properties": {
          "queryStatus": {
            "type": "string",
            "enum": [
              "complete",
              "partial"
            ]
          },
          "servedFrom": {
            "type": "string",
            "enum": [
              "upstream",
              "cache"
            ]
          },
          "fetchedAt": {
            "type": "string"
          },
          "cacheAgeMs": {
            "type": "integer"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "queryStatus",
          "servedFrom",
          "fetchedAt",
          "warnings"
        ],
        "additionalProperties": false
      },
      "SearchResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OccurrenceSummary"
            }
          },
          "range": {
            "$ref": "#/components/schemas/TimeRange"
          },
          "nextCursor": {
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "items",
          "range",
          "meta"
        ],
        "additionalProperties": false
      },
      "HistogramResponse": {
        "type": "object",
        "properties": {
          "range": {
            "$ref": "#/components/schemas/TimeRange"
          },
          "intervalMs": {
            "type": "integer"
          },
          "buckets": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "startUnixNano": {
                  "type": "string"
                },
                "endUnixNano": {
                  "type": "string"
                },
                "count": {
                  "type": "integer"
                }
              },
              "required": [
                "startUnixNano",
                "endUnixNano",
                "count"
              ],
              "additionalProperties": false
            }
          },
          "total": {
            "type": "integer"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "range",
          "intervalMs",
          "buckets",
          "total",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "Payload": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "available"
                ]
              },
              "value": {
                "$ref": "#/components/schemas/DebugEnvelopeV1"
              }
            },
            "required": [
              "status",
              "value"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "absent",
                  "invalid",
                  "unsupported_version"
                ]
              },
              "raw": {
                "type": "string"
              }
            },
            "required": [
              "status"
            ],
            "additionalProperties": false
          }
        ]
      },
      "OccurrenceDetail": {
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/OccurrenceSummary"
          },
          "exception": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "stacktrace": {
                "type": "string"
              },
              "mechanism": {
                "type": "string",
                "enum": [
                  "manual",
                  "window.error",
                  "unhandledrejection"
                ]
              },
              "handled": {
                "type": "boolean"
              },
              "location": {
                "type": "object",
                "properties": {
                  "url": {
                    "type": "string"
                  },
                  "line": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "column": {
                    "type": "integer",
                    "minimum": 0
                  }
                },
                "required": [],
                "additionalProperties": true
              }
            },
            "required": [],
            "additionalProperties": true
          },
          "payload": {
            "$ref": "#/components/schemas/Payload"
          },
          "storedFields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "summary",
          "exception",
          "payload",
          "storedFields",
          "warnings"
        ],
        "additionalProperties": false
      },
      "Capabilities": {
        "type": "object",
        "properties": {
          "apiVersion": {
            "type": "string"
          },
          "serverTimeUnixNano": {
            "$ref": "#/components/schemas/Wire_unixNano"
          },
          "filterFields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "service",
                "environment",
                "exceptionType",
                "message",
                "release",
                "traceId",
                "runtimeId",
                "groupKey",
                "origin"
              ]
            }
          },
          "filterOperators": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "and",
                "or",
                "not",
                "eq",
                "contains",
                "icontains",
                "prefix",
                "exists"
              ]
            }
          },
          "facetFields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "service",
                "environment",
                "release"
              ]
            }
          },
          "maxOffset": {
            "type": "integer"
          },
          "source": {
            "type": "string"
          },
          "features": {
            "type": "object",
            "properties": {
              "grouping": {
                "type": "boolean"
              },
              "snapshotDiff": {
                "type": "boolean"
              },
              "traces": {
                "type": "boolean"
              },
              "metrics": {
                "type": "boolean"
              },
              "queryFilters": {
                "type": "boolean"
              },
              "histogram": {
                "type": "boolean"
              },
              "alertExport": {
                "type": "boolean"
              },
              "investigation": {
                "type": "boolean"
              },
              "facets": {
                "type": "boolean"
              }
            },
            "required": [
              "grouping",
              "snapshotDiff",
              "traces",
              "metrics",
              "queryFilters",
              "histogram",
              "alertExport",
              "investigation",
              "facets"
            ],
            "additionalProperties": false
          },
          "maxPageSize": {
            "type": "integer"
          },
          "maxRangeMs": {
            "type": "integer"
          }
        },
        "required": [
          "apiVersion",
          "serverTimeUnixNano",
          "filterFields",
          "filterOperators",
          "facetFields",
          "maxOffset",
          "source",
          "features",
          "maxPageSize",
          "maxRangeMs"
        ],
        "additionalProperties": false
      },
      "RelatedRequest": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "same_span",
              "same_trace",
              "same_runtime",
              "time_window"
            ]
          },
          "range": {
            "$ref": "#/components/schemas/TimeRange"
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          }
        },
        "required": [
          "kind"
        ],
        "additionalProperties": false
      },
...
      "RelatedData": {
        "type": "object",
        "properties": {
          "evidence": {
            "$ref": "#/components/schemas/RelationEvidence"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelatedLog"
            }
          }
        },
        "required": [
          "evidence",
          "items"
        ],
        "additionalProperties": false
      },
...
      "TraceData": {
        "type": "object",
        "properties": {
          "traceId": {
            "type": "string"
          },
          "completeness": {
            "type": "string",
            "enum": [
              "unknown",
              "partial"
            ]
          },
          "spans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraceSpan"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "traceId",
          "completeness",
          "spans",
          "warnings"
        ],
        "additionalProperties": false
      },
      "RelatedResponse": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "available"
                ]
              },
              "data": {
                "$ref": "#/components/schemas/RelatedData"
              }
            },
            "required": [
              "status",
              "data"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "partial"
                ]
              },
              "data": {
                "$ref": "#/components/schemas/RelatedData"
              },
              "reason": {
                "type": "string"
              }
            },
            "required": [
              "status",
              "data",
              "reason"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "not_found",
                  "unavailable",
                  "not_configured"
                ]
              },
              "reason": {
                "type": "string"
              }
            },
            "required": [
              "status"
            ],
            "additionalProperties": false
          }
        ]
      },
      "TraceResponse": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "available"
                ]
              },
              "data": {
                "$ref": "#/components/schemas/TraceData"
              }
            },
            "required": [
              "status",
              "data"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "partial"
                ]
              },
              "data": {
                "$ref": "#/components/schemas/TraceData"
              },
              "reason": {
                "type": "string"
              }
            },
            "required": [
              "status",
              "data",
              "reason"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "not_found",
                  "unavailable",
                  "not_configured"
                ]
              },
              "reason": {
                "type": "string"
              }
            },
            "required": [
              "status"
            ],
            "additionalProperties": false
          }
        ]
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}